+++ src/test/converter2/issues/gh2916.js
@@ -1,10 +1,9 @@
 /**
- * @typedef {object} HelloProps
+ * @typedef {object} HelloProps {@inline}
  * @property {string} name Name property docs
  */
 
 /**
- * @inlineType HelloProps
  * @param {HelloProps} props
  */
 export function hello(props) {