+++ src/test/converter2/issues/gh3050.js
@@ -1,0 +1,13 @@
+/**
+ * @typedef Foo
+ * @property {number} a
+ * @property {boolean} b
+ */
+
+/**
+ * @typedef {Omit<Foo, "b">} Bar
+ * {@interface}
+ * {@interface with text}
+ */
+
+export {};